+++ src/renderer/src/components/recipes/RecipeEditor.jsx
@@ -164,6 +164,7 @@ function RecipeEditor({
         onFolderChange={handleFolderChange}
         onRegenerate={handleRegenerate}
         onSave={() => onSave(data, fileName, normalizeFolder(folder))}
+        onRenameFile={() => onSave(data, fileName, normalizeFolder(folder), 'rename')}
         onArchive={onArchive}
         onUnarchive={onUnarchive}
       />